requirements.txt
annotated-types==0.7.0
anyio==4.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.2
idna==3.10
joblib==1.4.2
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
ruff==0.11.7
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
